+++ src/components/Navbar.tsx
@@ -14,6 +14,9 @@ const Navbar = () => {
         <NavLink className={({ isActive }) => (isActive ? classes.active : classes.inactive)} to="/profile">
           Profile
         </NavLink>
+        <NavLink className={({ isActive }) => (isActive ? classes.active : classes.inactive)} to="/create">
+          Create
+        </NavLink>
         <button className={classes.login}>Login</button>
       </div>
     </nav>
