+++ src/pages/Create.module.css
@@ -1,0 +1,18 @@
+.container {
+  width: 250px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 30px;
+}
+.postForm {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: max-content;
+  height: auto;
+  background-color: #c8c8c8;
+  padding: 10px 8px;
+  border-radius: 6px;
+  gap: 5px;
+}
